% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textBoxPlacement.R
\name{axesRange}
\alias{axesRange}
\title{axesRange}
\usage{
axesRange(xList)
}
\arguments{
\item{xList}{list of numerical vectors containing axis points}
}
\value{
xlim
}
\description{
compute correct axes ranges for a set of overlain plots
}
\examples{
axesRange(yList1)
}
