% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textBoxPlacement.R
\name{xlim}
\alias{xlim}
\title{xlim}
\usage{
xlim(xList)
}
\arguments{
\item{xList}{list whose components are numeric vectors of the x values for overlaid curves}
}
\value{
numeric vector xlim
}
\description{
compute the numeric vector xlim for a set of curves
}
\examples{
xlim(xList1)
}
