% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textBoxPlacement.R
\name{xPosCheck}
\alias{xPosCheck}
\title{xPosCheck}
\usage{
xPosCheck(xPos, xList, verbose)
}
\arguments{
\item{xPos}{integer specifying x position to try to place text box}

\item{xList}{list whose components are numeric vectors of the x values for overlaid curves}

\item{verbose}{Boolean if TRUE print informative or diagnostic messages to console}
}
\value{
numeric vector valid values of xPos
}
\description{
is the value of xPos within a valid range?
}
\examples{
# replace incorrect xPos with reasonable value

xPosCheck(c(1,1,-5),xList2,verbose=TRUE)
}
