% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textBoxPlacement.R
\name{ylim}
\alias{ylim}
\title{ylim}
\usage{
ylim(yList, yrange, textBoxHeights, sortB, permInd)
}
\arguments{
\item{yList}{list whose components are numeric vectors of the y values for overlaid curves}

\item{yrange}{numeric vector (max-min) for vector of y values}

\item{textBoxHeights}{return value of textBoxUserUnits()}

\item{sortB}{Boolean if TRUE staggered curves are reordered, with
largest range curve on bottom of graph}

\item{permInd}{return value of permInd()}
}
\value{
returns a numeric vector ylim
}
\description{
compute the numeric vector ylim
}
\examples{
# demonstrate effect of sorting the curves

plot.new()
ylim(yList1,yrange(yList1),textBoxUserUnits(textList,yrange(yList1),verbose=TRUE),
  FALSE,permInd(yrange(yList1)))

}
