% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textBoxPlacement.R
\name{yrange}
\alias{yrange}
\title{yrange}
\usage{
yrange(yList)
}
\arguments{
\item{yList}{list whose components are numeric vectors of the y values for overlaid curves}
}
\value{
numeric vector yrange (max-min) for vector of y values
}
\description{
compute the staggered y values for the overlay plot
}
\examples{
yrange(yList1)
}
